Disable paranoid checks for now
Occasionally we see errors in compaction for non-increasing keys.
This is because RocksDB expects the sequence number to be in
decreasing order, but right now we follow the lexicographic order,
which is increasing.

Key layout is in db/memtable.cc:
[ key len | user key | packed sequence no. + type | value len | value ]

Proper fix would require us to copy the key and mutate it when
inserting, and keep the buffer in thread-local state. Since we're
not trying to take this to production, just disable paranoid checks
for now instead.